roll hardware_stats json and loop args into pio
rather than have two separate apps, add an update interval and json output
to pio.  clean up parseRunString() to always throw an exception on a
problem, rather than sometimes throwing and sometimes returning an error
code.

this modifies the json schema from hardware_stats: the error field is only
present if there is an error, and the value is a string describing the
error.  the output state is now on or off, same as the toString() method of
the OutputState class.  likewise button state is "up" or "down".  using the
string values rather than booleans seems more consistent with the standard
pio text output and more descriptive than just true or false.