add timestamp, add devices map to json
Add a timestamp field to json schema, and separate the status objects into
their own top-level "devices" object; the timestamp is not shown by default
in text output, but can be enabled with the --timestamp option.  Print an
error when the --update option is used outside of getting all status.
Allow specific devices to be passed into write_all_status() so the same
status output can be shared for one or all devices, and the json schema is
the same for both.

Encapsulate rts device access so status is always included, but since the
RTS setting does not persist after the fd is closed, I'm not sure this is
useful anyway.