preserve sensor/tag order in json and reporting
the map of sample id to SampleCounter does not need to be ordered, and so
it is now a std::unordered_map instead of std::map in case that gains some
efficiency.  meanwhile, a separate vector of SampleCounter* is kept to
match the xml order, and that vector is enumerated when reporting.  since
json does not guarantee key order in objects, and in fact jsoncpp seems to
sort keys lexically, the json schema now includes arrays "variable_list"
and "stream_list" with the variable name and key id in xml order.  see
NCAR/dsm-dashboard#74.

data_stats test baseline updated to pass again.